+++ CHANGELOG.md
@@ -116,6 +116,7 @@ and disable fields when "Settings (Inventory): Create, edit and delete HRID hand
 * Use `ResetProvider` to subscribe and publish the reset event and unsubscribe from it after a segment switch. Refs UIIN-3093.
 * Rename inventory.consortia permissions. Refs UIIN-3076.
 * Disallow displaying shared instances when find instance modal is open. Refs UIIN-3072.
+* ECS: Item affiliation cannot be changed when it is attached to a local order. Refs UIIN-3063.
 
 ## [11.0.5](https://github.com/folio-org/ui-inventory/tree/v11.0.5) (2024-08-29)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.4...v11.0.5)